On my machine the `parallelVerifyIsFasterThanSequentialVerify()` test failed constantly. I’ve increased the warm up cycle number, the count number and run the test several time using JVM8 on a Mac. Only after switching the order of sequential vs. parallel execution, the tests passed. Apparently the non-parallel stream did the warm up job or the parallel execution was actually faster.