@BloodyRookie 2016-07-07
Interesting, how many hardware threads does your computer have?
With 12 hardware threads (i7-5820k) i get without altering the test order:
INFORMATION: Parallel verify: 65095 ns/tx
INFORMATION: Sequential verify: 353432 ns/tx
So on my computer the ratio is 5.4