@loxal 2016-07-10
I have a Mac Pro with an i7 CPU and 8 hardware threads.