@BloodyRookie 2016-07-12
So in your commit, you don't have any warmup any more? That gives the first stream (not matter if sequential or parallel) always a disadvantage since the code has to be compiled and optimized in that run.