@loxal 2016-07-14
Well, I am not sure. Warm up is pretty JVM internal magic thingy and I am not sure whether the warmup knowledge applies in the same way for internal iteration of parallel & single streams as to classic external for-loop iteration. In fact the test passed with & without warmup and with swapping the execution order of the parallel with the single stream.