@BloodyRookie 2016-07-16
I think not only the warmup but also the garbage collection plays a role. The warmup is there to give java a chance to compile and optimize the crypto function verify().
I ran different combinations again. To see the influence of the garbage collection i also ran the test with default VM settings and a second time with enough RAM to hopefully avoid full garbage collections.
Here are the results:

Processor: I7-5820k @3.30GHz (6 cores, 12 hardware threads)
IDE: Intellij 2016.1.2

1) VM options: default

With warmup (parallel verify first):
------------------------------------
INFORMATION: Parallel verify: 52369 ns/tx
INFORMATION: Sequential verify: 402297 ns/tx
Ratio sequential/parallel: 7.68


Without warmup (parallel verify first):
---------------------------------------
INFORMATION: Parallel verify: 83330 ns/tx
INFORMATION: Sequential verify: 449851 ns/tx
Ratio sequential/parallel: 5.40


With warmup (sequential verify first):
--------------------------------------
INFORMATION: Sequential verify: 385270 ns/tx
INFORMATION: Parallel verify: 58983 ns/tx
Ratio sequential/parallel: 6.53


Without warmup (sequential verify first):
-----------------------------------------
INFORMATION: Sequential verify: 442512 ns/tx
INFORMATION: Parallel verify: 79388 ns/tx
Ratio sequential/parallel: 5.57


2) VM options: -Xms=12G, -Xmx=12G

With warmup (parallel verify first):
------------------------------------
INFORMATION: Parallel verify: 53549 ns/tx
INFORMATION: Sequential verify: 392899 ns/tx
Ratio sequential/parallel: 7.33


Without warmup (parallel verify first):
---------------------------------------
INFORMATION: Parallel verify: 66325 ns/tx
INFORMATION: Sequential verify: 399277 ns/tx
Ratio sequential/parallel: 6.02


With warmup (sequential verify first):
--------------------------------------
INFORMATION: Sequential verify: 389733 ns/tx
INFORMATION: Parallel verify: 55701 ns/tx
Ratio sequential/parallel: 7.00


Without warmup (sequential verify first):
-----------------------------------------
INFORMATION: Sequential verify: 406130 ns/tx
INFORMATION: Parallel verify: 56267 ns/tx
Ratio sequential/parallel: 7.21

As you can see the warmup and a huge heap tends to make the comparison more stable. I don't understand why you are experiencing different results. Could you do a test like the above for your machine?